Inherit root state when rendering a <Symbol /> component
The issue we were having:
When a symbol component is being created, it can reference the `bindings`
of the page it's on. However, when the symbol is actually inserted in
the page, the bindings can no longer be referenced.

This is because the root state of the application wasn't being passed
down to the symbol. We pass our custom bindings to the builder component
we render, and all child components use those bindings from the page's
root state. However, when the boundary between a page and a symbol is
crossed, those bindings weren't being passed down.